See Desc.
The thing that took the most time was figuring out how to have ImGui's floating windows call my callbacks correctly so that I could asynchronously poll the right data as well as use the event system. Due to limitations on the implementation however, asynchronous mouse position polling is completely inaccurate due to the fact that the position GLFW supplies is relative to a specific window. Again however, this I believe can be worked around by simply adding the main window's position to its relative mouse coordinates, and then for whatever window you want relative coordinates for, simply subtract its position from the absolute mouse position. This second step does not need to be performed for change in mouse position as the change will be the same regardless of where the origin is.